Only fetch brothers' photos

+++ get.py
@@ -14,6 +14,9 @@ headers = {
     'Cookie': cookie,
 }
 
+import json
+brothers = json.loads(open('/Users/erik/src/sigchibrothers/brothers.json', 'r').read())
+
 filename = 'page.html'
 if not os.path.exists(filename):
     print('Page not cached, fetching...')
@@ -74,7 +77,7 @@ for container in containers:
 
     image_id = container.find("div", {"class": "student_img"}).find("img")["src"][len("/facebook/Photo?id="):]
     image_filename = 'images/' + forename + ' ' + surname + '.jpg'
-    if not os.path.exists(image_filename):
+    if not os.path.exists(image_filename) and (forename + ' ' + surname) in brothers:
         image_r = requests.get('https://students.yale.edu/facebook/Photo?id=' + image_id,
                                headers=headers)
         with open(image_filename, 'wb') as f:
